pyproject.toml
[project]
name = "softmax"
version = "0.1.0"
description = "Code for 'Gradient Flow Polarizes Softmax Outputs towards Low-Entropy Solutions'"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "datasets>=3.0.0",
    "accelerate>=0.25.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "wandb>=0.16.0",
    "matplotlib>=3.10.8",
    "ipykernel>=7.1.0",
    "seaborn>=0.13.2",
    "jupyter>=1.1.1",
    "ipywidgets>=8.1.8",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
axlearn = ["axlearn[core]"]

[tool.uv.sources]
axlearn = { path = "../axlearn-mod", editable = true }
